requirements.txt
discord~=2.2.3
requests~=2.28.2
python-dotenv~=1.0.0
pyyaml~=6.0
Pillow~=9.5.0
urllib3~=1.26.16
aiohttp~=3.8.4
httpx~=0.24.1
inquirerpy~=0.3.4

# Language models
gpt4all~=0.2.1 # For GPT4All
imaginepy # For upscaling
gpt4free

# Only for Replit keep_alive
flask~=2.3.2
